app.py
import os
import boto3
import mysql.connector
from flask import Flask, request, jsonify
from tensorflow.keras.applications import ResNet50
from tensorflow.keras.layers import Dense, Flatten
from tensorflow.keras.models import Model
from transformers import GPT2LMHeadModel, GPT2Tokenizer

# Initialize Flask app
app = Flask(__name__)

# Function to connect to the relational database and retrieve data
def retrieve_data_from_database():
    # Connect to MySQL database
    conn = mysql.connector.connect(
        host=os.environ.get("DB_HOST"),
        user=os.environ.get("DB_USER"),
        password=os.environ.get("DB_PASSWORD"),
        database=os.environ.get("DB_NAME")
    )

    # Execute SQL query to retrieve data (example)
    cursor = conn.cursor(dictionary=True)
    cursor.execute("SELECT image_path, description, material, design_attributes FROM products")
    data = cursor.fetchall()

    # Close cursor and connection
    cursor.close()
    conn.close()

    return data

# Function to retrieve images from cloud storage
def retrieve_images_from_cloud_storage():
    # Initialize AWS S3 client
    s3 = boto3.client('s3', aws_access_key_id=os.environ.get("AWS_ACCESS_KEY_ID"), aws_secret_access_key=os.environ.get("AWS_SECRET_ACCESS_KEY"))

    # Example code to retrieve images from S3 bucket
    # Replace 'your-bucket-name' and 'your-prefix' with actual bucket name and prefix
    objects = s3.list_objects_v2(Bucket=os.environ.get("S3_BUCKET_NAME"), Prefix=os.environ.get("S3_PREFIX"))
    image_paths = [obj['Key'] for obj in objects['Contents']]

    return image_paths

# Function to preprocess images and generate features
def preprocess_images_and_generate_features(image_paths):
    # Initialize ResNet50 model
    base_model = ResNet50(weights='imagenet', include_top=False, input_shape=(224, 224, 3))

    # Add custom top layers
    x = base_model.output
    x = Flatten()(x)
    output_material = Dense(NUM_MATERIAL_CLASSES, activation='softmax', name='material')(x)
    output_design = Dense(NUM_DESIGN_CLASSES, activation='softmax', name='design_attributes')(x)

    # Combine base model with custom top layers
    model = Model(inputs=base_model.input, outputs=[output_material, output_design])

    # Process each image
    features = []
    for image_path in image_paths:
        # Preprocess image (resize, normalize, etc.)
        # Generate features using the model
        # Append features to the list
        features.append(...)  # Replace ... with actual feature generation logic

    return features

# Function to generate product descriptions using GPT-2
def generate_product_descriptions(data):
    tokenizer = GPT2Tokenizer.from_pretrained("gpt2")
    model = GPT2LMHeadModel.from_pretrained("gpt2")

    descriptions = []
    for item in data:
        # Generate description using GPT-2 based on item details
        description = "A beautiful {} {} with {} design.".format(item['material'], item['product_type'], item['design_attributes'])
        descriptions.append(description)

    return descriptions

# Define API endpoint for generating product information
@app.route('/generate_product_information', methods=['POST'])
def generate_product_information():
    # Retrieve data from database
    data = retrieve_data_from_database()

    # Retrieve images from cloud storage
    image_paths = retrieve_images_from_cloud_storage()

    # Preprocess images and generate features
    features = preprocess_images_and_generate_features(image_paths)

    # Generate product descriptions
    descriptions = generate_product_descriptions(data)

    # Combine data and descriptions
    product_information = []
    for i, item in enumerate(data):
        product_info = {
            'image_path': image_paths[i],
            'description': descriptions[i],
            'material': item['material'],
            'design_attributes': item['design_attributes']
        }
        product_information.append(product_info)

    return jsonify(product_information)

# Main function to run the Flask app
if __name__ == '__main__':
    app.run(debug=True)
