requirements.txt
tensorflow
mysql-connector-python
pandas
Flask
boto3
transformers